Add files via upload
Making graph by 2 approaches (adjacency list and adjacent matrix.). The matrix is built with help of a 2d array while the list is built with help of vectors because we need an array of linked lists. So vectors are ideal for that matter. All functions in the directional and undirectional way are included like display, add edge and remove the edge.